refactor: move database connection to backend root folder
